+++ src/app/constants/index.ts
@@ -1,6 +1,7 @@
 export * from "./error-messages.constants";
 export * from "./environment.constants";
 export * from "./fail-open.constants";
+export * from "./mdn-locale.constants";
 export * from "./rate-limit.constants";
 export * from "./react-translation.constants";
 export * from "./runtime.constants";